+++ src/pause/input_menu.rs
@@ -83,7 +83,7 @@ impl PauseMenu {
                 Vector2::new(item_x, y),
                 size,
                 0.0,
-                palette::color(Pal::White),
+                palette::engine_color(Pal::White),
             );
             if i == self.input_menu_selected {
                 draw_indicator(d, self.time, item_x, y + size * 0.5);